Add new constraints.
New constraints on the angular acceleration that take the place of the
"mu w h" constraint we had previously. The number of iterations for the
SQP was increased from 3 to 5, as it otherwise failed to satisfy the
constraints sometimes.

Tested with objects that violate the "mu w h" constraint, and it does
work. It seems to make x acceleration lower, which is the expected
result.